+++ tests/Utils/Enums/ImageableType.php
@@ -1,8 +1,10 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Tests\Utils\Enums;
 
+use Tests\Utils\Models\Task;
+
 enum ImageableType: string
 {
-    case TASK = \Tests\Utils\Models\Task::class;
+    case TASK = Task::class;
 }